Revert "Read as integer instead of long"
This reverts commit 6300262d607b305ae55b8edf5bf787a7cdf7eeaa.

+++ proxy/src/main/java/com/velocitypowered/proxy/config/VelocityConfiguration.java
@@ -763,8 +763,8 @@ public final class VelocityConfiguration implements ProxyConfig {
       final boolean kickExisting = config.getOrElse("kick-existing-players", false);
       final boolean enablePlayerAddressLogging = config.getOrElse(
               "enable-player-address-logging", true);
-      final int maxCommandsPerSecond = config.getOrElse(
-              "max-commands-per-second", 10);
+      final long maxCommandsPerSecond = config.getOrElse(
+              "max-commands-per-second", 10L);
       final boolean logPlayerConnections = config.getOrElse(
               "log-player-connections", true);
       final boolean logPlayerDisconnections = config.getOrElse(
